backend/requirements.txt
Flask
gunicorn
Flask-Cors
Flask-JWT-Extended
python-dotenv
requests
Werkzeug
mysql-connector-python
mysql
dnspython
pymongo